Move the window management functionality to a separate page.
